Handle directory traversal attacks in safe_join
Add dependency on `sugar_path` for the `normalize` method, which removes
`..` without the final path needing to exist. We require this to match
the Django algorithm, which allows returning a path that doesn't exist.
In `FileSystemLoader` we rely on `safe_join` to return paths that don't
exist so we can report them in `LoaderError.tried`.