Refactor Filter and introduce Render trait
All `Filter`s have a location `at` and a `left`, but they vary on
whether they take a `right` argument and whether this is optional or
required. Now this information is stored in `FilterType` for each
variant.

Introduce a `Render` trait to define an interface for resolving a
template element to a form that can be rendered or passed to another
element. Also add a `render` method to convert a resolved element into a
string.