Start passing around &EngineData
This struct contains data set on the `Engine` that the parser and
renderer need access to to work correctly. We can't pass a reference to
the full `&Engine` because we need mutable access to `template_loaders`
at the same time.

Add `autoescape` and `libraries` to `EngineData` for now. `autoescape`
is already used and `libraries` will be used when parsing the
`{% load %}` template tag.