Test more center filter edge-cases at isize limits
Python's `center` function (which Django uses in the `center` filter)
overflows for integer arguments outside of C's `ssize_t`. On a 64 bit
system, this corresponds to the range `(-2**63 - 1, 2**63)`.

We can represent this range in Rust with the `isize` type, which lines
up neatly with the maximum capacity of a Rust `String` (which is backed
by a `Vec`):

https://doc.rust-lang.org/std/vec/struct.Vec.html#panics