Work around badly behaved custom tags cloning the context
Django's custom tags (for example defined with `register.simple_tag`)
can be passed the full `context` object to introspect. Most
implementations are well behaved and only read and write to the
`context` object. It is possible, however, for a tag to keep a reference
to the `context` around (for example using the `global` keyword).

This causes a problem for the Rust implementation which wants to take
back ownership of the `context`. In this commit, we optimistically try
to take ownership of the `context` back from the `PyContext` wrapper
type, using `Arc::try_unwrap`. For badly behaved tags, this fails and we
fall back to cloning the Rust `context` object.

This allows the Rust code to keep going with a correctly structured
`context` while leaving Python with a valid `PyContext` instance.